Remove error sender trait, now that there is only one impl
